@@ -2,11 +2,12 @@ package migrate
 
 import (
 	"fmt"
+	"os"
+	"slices"
+
 	"github.com/packwiz/packwiz/cmdshared"
 	"github.com/packwiz/packwiz/core"
 	"github.com/spf13/cobra"
-	"golang.org/x/exp/slices"
-	"os"
 )
 
 var loaderCommand = &cobra.Command{
